And fixing up the UniqueId test cases to end the night.

+++ lib/Dht/UniqueId.cpp
@@ -6,7 +6,7 @@ namespace IsoPeer { namespace Substrate { namespace Dht {
     UniqueId UniqueId::Distance(const UniqueId& first, const UniqueId& second)
     {
         UniqueId returnId;
-        for (int i = 0; i < first.m_bytes.size(); i++)
+        for (size_t i = 0; i < first.m_bytes.size(); i++)
         {
             returnId.m_bytes[i] = first.m_bytes[i] ^ second.m_bytes[i];
         }
